requirements.txt
Flask==2.3.3
twilio==8.1.0
openai==0.28.0
reportlab==4.0.4
google-cloud-storage==2.10.0
python-dotenv==1.0.0
gunicorn==21.2.0
firebase-admin==6.2.0